+++ core/detections/list.js
@@ -74,6 +74,11 @@ const Converter = require('../../common/converter')
  *     responses:
  *       200:
  *         description: List of detection (lite) objects
+ *         headers:
+ *           Total-Items:
+ *             schema:
+ *               type: integer
+ *             description: Total number of items without limit and offset.
  *         content:
  *           application/json:
  *             schema:
@@ -94,8 +99,7 @@ router.get('/', (req, res) => {
   converter.convert('classifiers').optional().toArray()
   converter.convert('classifier_jobs').optional().toArray()
   converter.convert('min_confidence').optional().toFloat()
-  // TODO: improve validation tool to validate possible values for array items
-  converter.convert('review_statuses').optional().toArray()
+  converter.convert('review_statuses').optional().toArray().isEqualToAny(['unreviewed', 'rejected', 'uncertain', 'confirmed'])
   converter.convert('limit').optional().toInt().maximum(1000)
   converter.convert('offset').optional().toInt()
   converter.convert('descending').default(false).toBoolean()
@@ -112,7 +116,7 @@ router.get('/', (req, res) => {
         user
       }
       const result = await dao.query(filters, options)
-      return res.json(result)
+      return res.header('Total-Items', result.total).json(result.results)
     })
     .catch(httpErrorHandler(req, res, 'Failed getting detections'))
 })