+++ core/streams/create.js
@@ -1,9 +1,6 @@
 const { httpErrorHandler } = require('../../common/error-handling/http')
-const dao = require('./dao')
-const { randomId } = require('../../common/crypto/random')
 const Converter = require('../../common/converter')
-const arbimonService = require('../_services/arbimon')
-const { ValidationError } = require('../../common/error-handling/errors')
+const { create } = require('./bl')
 
 /**
  * @swagger
@@ -36,6 +33,10 @@ const { ValidationError } = require('../../common/error-handling/errors')
  */
 module.exports = (req, res) => {
   const user = req.rfcx.auth_token_info
+  const creatableBy = user.is_super || user.has_system_role ? undefined : user.id
+  const requestSource = req.headers.source
+  const idToken = req.headers.authorization
+  const options = { creatableBy, requestSource, idToken }
   const converter = new Converter(req.body, {}, true)
   converter.convert('id').optional().toString().minLength(12).maxLength(12).isPassingRegExp(/[a-z0-9]{12}/, 'should have only lowercase characters and integers')
   converter.convert('name').toString()
@@ -50,46 +51,9 @@ module.exports = (req, res) => {
 
   return converter.validate()
     .then(async (params) => {
-      const stream = {
-        ...params,
-        createdById: user.id
-      }
-      if (!params.id) {
-        stream.id = randomId()
-      }
-
-      if (params.latitude === 0) {
-        stream.latitude = null
-      }
-
-      if (params.longitude === 0) {
-        stream.longitude = null
-      }
-
-      if (params.projectId) {
-        const duplicateStreamInProject = await dao.query({ names: [params.name], projects: [params.projectId] }, { fields: 'id' })
-        if (duplicateStreamInProject.total > 0) {
-          throw new ValidationError('Duplicate stream name in the project')
-        }
-      }
-
-      // TODO move - route handler should not contain business logic
-      if (arbimonService.isEnabled && req.headers.source !== 'arbimon') {
-        try {
-          const externalSite = await arbimonService.createSite(stream, req.headers.authorization)
-          stream.externalId = externalSite.site_id
-        } catch (error) {
-          console.error(`Error creating site in Arbimon (stream: ${stream.id})`)
-          throw new Error()
-        }
-      }
-
-      const options = {
-        creatableBy: (user.is_super || user.has_system_role) ? undefined : user.id
-      }
-      const createdStream = await dao.create(stream, options)
-
-      return res.location(`/streams/${createdStream.id}`).sendStatus(201)
+      params.createdById = user.id
+      return await create(params, options)
     })
+    .then(stream => res.location(`/streams/${stream.id}`).sendStatus(201))
     .catch(httpErrorHandler(req, res, 'Failed creating stream'))
 }